refactor: 리듀서 분리
- itemListReducer: 상품 목록 관련
- orderListReducer: 주문 내역 관련
