+++ src/store/index.js
@@ -1,46 +1,9 @@
 import { combineReducers, createStore } from 'redux';
 import { composeWithDevTools } from 'redux-devtools-extension';
-import { ACTION_TYPE } from '../constants';
 import { cartReducer } from './cartReducer';
+import { itemListReducer } from './itemListReducer';
+import { orderListReducer } from './orderListReducer';
 
-const initialState = {
-  itemList: [],
-  orderList: [],
-};
-
-export const setItemList = data => {
-  return {
-    type: ACTION_TYPE.SET_ITEM_LIST,
-    payload: data,
-  };
-};
-
-export const addOrderDetail = data => {
-  return {
-    type: ACTION_TYPE.ADD_ORDER_DETAIL,
-    payload: data,
-  };
-};
-
-export const reducer = (state = initialState, action) => {
-  switch (action.type) {
-    case ACTION_TYPE.SET_ITEM_LIST:
-      return {
-        ...state,
-        itemList: action.payload,
-      };
-
-    case ACTION_TYPE.ADD_ORDER_DETAIL:
-      return {
-        ...state,
-        orderList: state.orderList.concat(action.payload),
-      };
-
-    default:
-      return state;
-  }
-};
-
-const rootReducer = combineReducers({ cartReducer, reducer });
+const rootReducer = combineReducers({ itemListReducer, cartReducer, orderListReducer });
 
 export const store = createStore(rootReducer, composeWithDevTools());
